requirements.txt
Flask
numpy
opencv-python-headless
requests
python-dotenv
google-generativeai
Pillow
gunicorn